Add files via upload
I've had a technical issue with numpy and pandas libraries not working as they should on my pi 400 so I haven't properly tested this program yet but it's constituent parts have all been tested and are known to work smothly without any issues.  If someone else out there wants to test this program for me please do feel welcome to run it on your own machine.